up: workaround OSX issues
Apparently `docker-machine` cannot be run under sudo on OSX. Rework the
script to require sudo only when needed. It's still bad to run it under
the hood, but there does not seem to be any other way.

Drop stat as format options are completely different. Instead create a
temporary file with hosts to be added that tee it into /etc/hosts.

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>