storage-proxy: storage proxy with rate and connection limiting
Add a storage proxy that applies download rate and connection limiting to
incoming GET requests.

The service is implemented using openresty and Lua extensions. When connection
limit is hit, a 503 Service Temporarily Unavailable is returned.

Minio is hidden behing storage proxy for all requests, i.e. internal requests
from Deployments Service will also go through storage proxy. This is required
because of signed request in S3, which the target host name is a part of.

The default limits are:
- download rate: 1MB/s
- max connections: 100

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>