Integrate docker-compose into test framework itself.
The motivation for this is to have different setups for different
tests. Most will use the standard setup with one bootstrapped client,
but other setups involving more clients are possible.

To do this, we have to make docker-compose part of the py.test
fixtures. Python was annoyingly strict about the order of loading
things, so I had to split it up quite a bit, into this dependency:

common
  ^
  |
  |\----------------+
  |                 |
common_docker   MenderAPI
  ^                 ^
  |                 |
  |/----------------+
  |
common_setup

Some other things changed:

* User authentication now has to be done after every docker-compose
  restart, so this has been moved to its own authentication module.

* We are now using docker-compose to query for node IPs.

* Tests are not order dependent anymore, so run them in default order.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>