compose/demo: add mongo data volume mounts
we already have them in prod.yml, but it's useful to also have them in demo -
the user might be trying out the system and recreating containers along the
way. it would be unexpected to lose whatever data was already there.

for simplicity using host mounted volumes (not external, manually defined ones like in prod).

Issues: MEN-879

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>