Make it an error to provide build dir that doesn't contain what we need
It's important not to fall back to downloading in this case, because
this can lead to subtle and difficult-to-reproduce errors because the
version differs from what is expected.

With this patch we grab everything from the build if the BUILDDIR is
specified, and we fail if this does not work. Only if the variable is
not set do we resort to downloading. In both cases we expect to find
executables in the path, rather than in the current directory.

Also do some of the preparation steps regardless of whether the files
already exist. It's not an expensive operation, and it's better to be
sure they are up to date.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>