Fix incorrect writing of file into ext4 using debugfs.
Debugfs really deserves its name, "debug": Writing into a file name,
"/etc/hosts", will create a literal file with that name, including
slashes and everything, in the root! The slash is not a valid
filesystem character, so this corrupts the filesystem, and certainly
the file won't be found in the correct place.

Work around this by using the debugfs "cd" feature to get to the
correct directory first. Unfortunately you can't pass more than one -R
parameter, so we need to pipe the commands into it instead.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>