migration: compose for mongo migration helper service
Add a compose file for a mongo instance that's used during migration. This is
needed for two reasons. First, DB must be dumped and restored by a
mongodump/mongorestore tools in the same version as the server and second, we'll
use compose to start the service so that it works with other backend
services (hence avoid second guessing compose naming idiosyncrasies).

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>

Changelog: None