migration/dump-1.0-db: helper for dumping 1.0 DB contents
The 1.0 release is running at least 5 distinct mongo nodes, each for one
service. If we want to migrate to using a single mongo instance in on-prem
setup, the data needs to be migrated.

This tool allows for dumping DB contents into a directory named 'db-dump'
located at the root of integration repository. Before dumping DB, all services
that use mongo are stopped.

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>

Changelog: None