migration/restore-db: helper for restoring a mongodump'ed DB
This is a helper script that can be used to a DB dump created with mongodump
tool. The helper will start 'mongo-helper' service, mounting the root of
integration repository into the contained. DB data is expected to be located in
'db-dump' directory, at the root of integration repo.

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>

Changelog: None
Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>