Add release_tool.py.
This script is meant to aid in the release process. As we need to
submit this is in order to test it during a release, it doesn't carry
much documentation at the moment, but this will come in a later
commit.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>
(cherry picked from commit d77818e53d1ab4dea7fe919e37838ec785d93ac8)