release_tool: Add support for merging release tag into version branch.
This is to ensure that a default clone will contain all important
tags, since by default Git doesn't fetch tags that don't point into
known branches.

This required a change in logic: We now keep following branches even
for repositories that are based on an already released tag. This is
because we need this information when doing this merging. However, we
still print that we follow the tag, since this is less confusing when
considering where the next build tag will come from (a repo based on a
released tag will never get a new build tag).

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>