release_tool: Don't use HEAD for local branches, but instead actual name.
HEAD doesn't work well when we try to clone a local repository for
doing changes, since it depends on a named branch.

For this reason also take into account whether we should track the
branch or not (affects if Git creates the branch from a remote or
not).

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>