Fix slightly incorrect image upload request, partic. content-type.
The existing code would submit the image file with content-type of
form-data, which, although the server will in fact accept it, is
obviously wrong. Form-data is implied simply by submitting data to a
post request, and covers the whole request, not just the image blob,
so this should not be specified.

Set the type of the image to application/octet-stream, which is what
it should be, and make the call a bit more readable.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>