conductor: improve loading of task and workflow definitions
Replace curl loader with an python script that does the same and more.

Task definitions will always be uploaded to conductor. Workflows are handled
with more care now, by actually looking at workflow name and version. Before
loading a workflow, we check with conductor if a workflow with identical name
and version already exists. Workflow is uploaded only if there is no match,
otherwise a message is logged.

Fixes: [Mender #1237]
Changelog: None
Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>