tests/run.sh: whitelist all IPs/DNS names in the gateway for tests
tests access the gateway via IP, not a well-defined alias like docker.mender.io.
now that we have a host whitelist (server_name: ...) in the gateway, we have to
set it - so set it to a catchall regex because we don't know anything in advance.
otherwise pretty much everything fails with a nasty connection termination error.

btw '*' doesn't work here, yaml doesn't allow it.

Issues: MEN-1262

Changelog: None

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>