tests: Switch to Docker teardown that doesn't depend on yml files.
The problem is that the tests use a myriad of different yml files, and
it's difficult to know when we are tearing down the environment, which
ones they used. So switch to pure docker command instead, and simply
remove everything in the current namespace, regardless of which yml
file originally launched them. This also has the advantage of being
faster, since "docker rm -f" acts immediately.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>