tests: Rewrite reboot detector to be notify instead of polling based.
This improves performance drastically, since we don't have to connect
every second to check the status, which, because of the SSH handshake,
is quite expensive in an emulated environment.

The downside is that it depends on the mender-reboot-detector service,
which must be running in the demo image. We store a special file on
the data partition to signal that the service should be activated,
otherwise it will just be dormant, so it won't mess with the demo
images that users run.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>