release_tool: Fix several problems when using -b option with branches.
Basically, since the tool expects to work with tags (which we should
when doing releasing), we need some extra logic to populate the
"build_tag" elements (though not really tags anymore) with branch
names instead, if that is what's requested. We leave the release part
alone, since it is important that tags are used there.

Also add ability to specify more than one pull request, and to specify
pull requests for meta-mender.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>