demo: suppress warning on newer docker-compose versions
docker-compose version 1.21.0 and greater has deprecated
the --parallel option printing the following if used:

    WARNING: --parallel option is deprecated and will be
    removed in future versions.

This is because "--parallel" is the default mode as of
version 1.21.0.

Suppress warning by only using the --parallel option
if --no-parallel option does not exist because this
was added together with deprecation of --parallel.

Changelog: Title

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>