release_tool: Handle missing repo in version range.
If you call --in-integration-version with a version range,
release_tool will also return a range for the queried repository. But
this broke if the given repository didn't exist in the older
version. This commit fixes it by simply returning the entire history
in this case. This makes sense since the repository is new.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit a33549bdfb6eb57c967222b278c1c52b7dd2d6b7)