release_tool: Fix tag reference error when not inspecting newest tag.
When we verify tag references (--verify-integration-references), we
only checked the most recent tag, because that is what `git describe`
returns. However, we need to check all tags to be able to build
multiple build tags in parallel that point to the same commit. For
this we use `git tag --points-at` instead and reorder the code a bit
to support a list of tags.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>