release_tool: Dynamically import Jenkins params instead of hardcoding.
This fixes the problem we face every time a Jenkins parameter is
added, that the job triggering stops working because that parameter is
missing from the list.

It also has the nice benefit that when changing defaults in Jenkins,
the defaults will be reflected in the build parameters that the
release_tool presents.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>