QA-37: Fix inability to query repos that have been removed.
We need to be able to query repos that have been in the product
earlier, but which have been removed, both when using the `--list`
argument and the `--version-of` argument. To achieve this we turn the
existing static list into a candidate map instead, and fill it based
on which versions are available either in the current working
directory or, if given the `--in-integration-version` argument, that
version.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>