Remove many state script tests and adjust remaining ones.
We keep a few, but this is already well covered in the new client unit
tests, and those tests are *much* faster, so we save quite a lot of
testing time by doing this.

Also fix a few state changes that have changed behavior after
rewriting the state machinery.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>