compose/demo: add demo version of minio's nginx config
add a demo config with max-age=0, use it in the 'demo' layer.

to avoid conflicts - don't mount any config in the 'storage.minio' layer.
it's then up to compose, based on the order of included files, to pick
one or another. demo + storage could give random results.

instead, add the existing, production config to the 'prod' template.

also, use the demo config in the 'testing' layer, as this duplicates the
demo layer for testing.

https://tracker.mender.io/browse/MEN-2440

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>