Remove timing dependent test which is tested in mender repo.
The test is highly timing dependent, since whether the client is
killed in the Download or the ArtifactInstall state decides whether
the resulting image name will be `installed_yocto_id` as expected, or
something with `INCONSISTENT`.

In addition this is now tested much better in the
`TestStateTransitionsWithUpdateModules` test in the mender client
repo.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>