Unify demo scripts by removing all except the one called 'demo'.
We don't actually need the `/etc/hosts` meddling which exists in the
`up` script, since Docker works just fine with its own internal DNS
server. The other scripts provide no extra functionality and are
completely redundant.

The only thing we os keep the `verify-docker-versions`, which was not
in `demo` before.

MEN-2571

Changelog: Title

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>