MEN-1939: Add request retrying for API calls.
This is motivated by the extremely unpredictable issue where MongoDB
will either return I/O timeout or connection refused, both of which
seem to be caused by high load. It is a frequently occurring issue in
our parallel tests, but when testing manually the issue always seems
to go away on subsequent attempts.

Hence this fix to add a retry mechanism when doing API requests and
the server returns 500 Server Error (and only then).

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>