tests: fix skip for test_00_multi_tenancy
the condition:
- is not available anymore
- was always True
- didn't make sense at all

the --mt_... arg was never provided. but - it always
had a default. so, going by this condition alone, the
test would always run, hence the 'makes no sense'.

what actually controls MT tests is the run_tenant_tests flag,
but the mechanism is very indirect.

MT fixtures in setup check this flag and do a pytest.skip.
what this function does is it *skips the entire test* the fixture
is called for. so this, if anything could be the condition for the test.

but then - this flag is also set to true by default. the only time
it's set to False is when we're in an integration branch, outside of
master.

conclusion:
- don't set any condition
- let the master vs branch logic set run_tenant_tests
- keep relying on fixtures to skip out of MT tests according to ^^

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>