Support trigger builds on GitLab from release_tool.py
The tool, by default, will still trigger the builds on Jenkins CI, while
this behaviour is now configurable using -c cli parameter.

The implementation basically follows the same logic as for Jenkins, only
deciding in which server to trigger the build in the last stage. One
notably difference is that the default params are fetched from the API
for Jenkins but parsed from .gitlab-ci.yaml for GitLab.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>