Modify "expect assert" logic for a less agresive one
Old code used assert as part of an expected logic path to just retry the
operation. This seems a bit of an overkill and has a undesirable side
effect: the logger messages are lost when the code jumps out asserting.

Also, let's be honest, it is lazy coding.

The current change is logic compatible with the old one. There are still
some assertions inside Authentication.get_auth_token() that could break
this logic, but such function is meant to be called once and not
expected to require retries.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>