release_tool: Improve `-f` to take `release_component` into account.
We only want `-f` to give results for components that have been part
part of a release. Those that haven't can have any arbitrary version
listed in the docker-compose files, and are not really *part* of a
release.

For this we have to do quite some restructuring: We need to move the
COMPONENT_MAPS out of the tool and into its own data file, so that we
can query it from other branches. Welcome component-maps.yml! We use
this to query for whether the component was releaseable in the version
that is being requested, with some special logic to handle cases where
the file is not present (old branches).

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit d86c7ff169fe9bd67a47443e871a9ae6fba59804)