Add Enterprise production template and restructure a little.
The restructuring is mainly to address a weakness in the original
template, which was that since the user was expected to copy all the
files in the template directory, it's not possible for us to provide
updates to the `run` script.

Addressed this by instead providing an OOTB production directory, with
tracked `run` script and templates named with `.template` suffix. This
should be completely compatible with existing installations except
that the user need to delete the `run` script from his repository.
This also means that we can cut the step in the production
guide to replace path names, because the files are already in the
correct directory.

Changelog: Introduced `enterprise.yml` template in production
directory to install an Enterprise backend server.

Changelog: The old `template` directory has been replaced with a
dedicated `production` directory, and templates are now provided as
single files with the `.template` suffix instead. These should be
copied to their non-`.template` location before being used. The `run`
script should no longer be copied, and if it already exists in the
`production` directory before merging this change, it should be
removed before attempting to merge or rebase.

Changelog: The `prod.yml` file has been moved into a `config`
subfolder. Users with downstream repositories need to move their
`prod.yml` as well.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>