tests/common_docker: exclude selected containers from stop_docker_compose
a new helper which will be useful when testing OS->ENT migration.

we might want to keep the mongo container around, so have a version
of the 'stop_...' helper (which actually destroys the whole
setup) that supports exclusions.

the command is ugly because of the inflexibility of filtering
via 'docker -f' - no way around it, grepping and awking is
the only choice.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>