testutils/infra: >1 filters on getid
we used -f name=, but it allows only a single filter on a given
attribute.

in integration, we'd want to filter both on service name, and
on the unique project prefix. each integration setup gets its
own.

to enable that - move to multiple filters via chained greps + awk.

CLIs must now accept an optional docker prefix to be usable
from both integration and backend-tests. same goes for common
helpers that use the clis.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>