compose: include mongo-tenantadm alias
must be defined also here, in the basic compose file.
we'll be doing a selective restart in the enterprise migration test -
mongo is left alive between runs.
if the alias is not there right from the start, tenantadm
started only in the second setup will never see its db.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>