Fix logic detect container exit code
The exit code of the docker inspect subshell will always be success, so
failed could not be set to 1 there (it was set to 1 anyway as part of
the docker attach part).

Anyway, modify a bit the logic to make the whole wrapper exit with the
same exit code than the container (and therefore the pytest call).

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>