Set pytest to v2.x to fix strange error in backend integ tests
There seems to be a modification in the way the Exceptions are expected
and reported in pytest from v3 onward. The correct fix is to modify the
test, but meanwhile...

The failure was introduced in d8c4c6b, when pytest was indirectly upgraded
to latest, but has been hidden due to the development of the test
happening in an isolated feature branch.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>