Add changelog tag for previous commit 2e51ab0
As this is a problem that affects the beta release, we need a Changelog
tag to inform users.

Changelog: Add enterprise enabling flag to enterprise composition GUI
container, so that the enterprise features are shown in the Frontend.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>
(cherry picked from commit 2c94262ebf81e1d4163d3d36add4cb21986886be)