MEN-2836: Fix - Make sure the demo-script subprocess has a stdin fd
The demo script, which was started from pytest did not have a open
stdin file-descriptor, which in turn caused the interactive read
call in the shell to never block, like it is supposed to. The fix
was giving the stdin of the subprocess a filedescriptor to block on.

Changelog: Title

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>