MEN-2836: Fix - Create explicit exitcond for the demo setup fixture
Previously the fixture which runs the demo script did read one line
at a time from the STDOUT of the demo script until it found the
'Login password: ' line, and then returned from the fixture with
the process handle for the test. Since running the demo script, then
stopping it, and restarting it persists the environment (and therefore
the user), the exit condition will be wrong, and hence the second time
the environment is brought up, the setup fixture has to exit on
the condition that the user already exists.

Changelog: Title

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>