Remove AWS S3 tests, fixtures and docker-compose file
These tests are not run in CI. The bucket specified in the file is
actually empty, so AFAIK they might have never been run.

Furthermore, the value of running few tests on AWS S3 storage is really
little, keeping in mind that we extensively test MinIO (same API).

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>