Remove nightly flag and fix/enable test_image_download_retry_hosts_broken
This test takes very long to run (around 8 minutes on my machine) so it
was set for "nightly" builds, which we don't have in place in our CI.
Keeping the test non executed has the obvious risk of not being updated,
so the test was actually not passing anymore.

Remove the nightly flag and promote this test to "slow".

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>