MEN-1940: Added the ability to run integration tests as a one-liner
Added a README file to the tests subdirectory, describing how to add the
required dependencies for the integration tests, and then running them.

The description goes through running the tests

* Locally - With no extra configuration
* Locally - With a Python virtual environment

The requirements are organized in the `tests/requirements` subfolder, with the
following directory structure:

tests/requirements
 |
  - python-requirements.txt
  - deb-requirements.txt
  - apt-requirements.txt

The README includes two examples.

* One for running the test with a custom backend service
* One for running the tests with a modified mender-client

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>