Proof of concept for encaptulating docker-compose layer
Very much open for discussion at this point.

This commit adds a new module container_manager where to hide the
interactions with containers. Partially moved docker-compose code to its
class and created a factory to showcase the design.

The idea is that different fixtures now can setup/teardown namespaces
(understood as a group of containers) in which to run tests, without
really caring if they are orchestrated by docker-compose, or minikube or
somewhat manually.

Many things are still missing, and at this points the tests use both
functionality from the new and the old files.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>