Refactor cli/docker testutils to reuse extending container_manger API
The motivation behind the refactor is to be able to reuse Cli* clases
with the new container_manager module, so that they become docker
agnostic.

At the moment the constructor accepts either a namespace string (for
which assumes docker, to be able to play nicely with backend-tests) or
an already created object obeying BaseContainerManagerNamespace
interface.

On the other hand, we move the plain docker wrapper to a docker_manager
where to implement a simpler ContainerManager that deals directly with
docker (ignores docker-compose as it asumes it is done outside of the
test framework).

To accomodate for this functionality, we have also extended base class
to use a name in constructor, which (almost accidentaly) allow us to get
the rid of the hack of setting a unique docker_compose_instance from
common_setup, passing it in the constructor instead.

Along the way, rename docker_prefix to a generic containers_namespace
(that was already the actual use of the parameter) and add the _1 suffix
to futher filter the docker ps output.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>