Move ssh_is_opened to Helpers class and remove default handling
This function has nothing to do with docker setup and it fits better
together with the rest of device facing funtions in the Helpers class.

In addition, we remove the magic handling for a default host value, so
that the caller now must pass the hosts for which to expect SSH
connection.

This commit breaks common_setup and several tests, but we will deal with
that shortly :)

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>