Decouple fixtures and tests from containers handling
Implement the remaining methods on docker_compose_manager to have the
same funtionality as the (now deprecated!) common_docker.

Then, make the fixtures return a namespace which encapsulates the docker
magic in a known interface. Fixed also the ssh_is_open calls.

Finally, modify test_basic_integration tests as a demo about the
intented behaviour of the tests: get the namespace through the fixture,
and use it to get mender clients IPs or other containers dependent
information instead of the old global functions.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>