Decouple now MenderAPI from containers handling
This commit makes MenderAPI class to require a container_manager
parameter on reset, offering a getter to the rest of the API classes
from where to inquiry the container related information (namely gateway
ip and mender clients).

The fixutres can now set the manager on each reset.

This is not the nicest design, but we will tackle the MenderAPI refactor
in a future task.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>

amend-MenderAPI-ugly