[tests/run.sh] Move DEFAULT_TESTS to use pytests' testpaths
This change will use pytests' 'testpaths' variable to set the default test
directory in which to run tests, hereby simplifying the 'run.sh' script by
removing the logic to handle overriding the default tests case.

Overriding the default tests files is still possibly, simply through giving it
as an argument to the script, so no functionality has changed.

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>