[tests/run.sh] Enable passing on quoted arguments to pytest
Previously the only way of running specific integration tests was to set the
$SPECIFIC_INTEGRATION_TEST variable to whatever regex that you wanted to run.

With the change, now this can be done straight from the command line through the
regular GNU convention of separating the arguments passed to the script itself,
from the arguments passed along to pytest through the ' -- ' separator.

Hence, now, one can run the script, with no dependency downloads, and only
running the enterprise tests like so:

$ run.sh --no-download -- -k 'not Enterprise'

For further information, see the '--help' command.

Changelog: Title

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>