Fix integration test failures: deployments unavailable
Integration tests uploading an artifact are just too fast and perform the API
call to the deployments micro-service before the minio service is up and
running. While these calls have a retry mechanism in place, the deployments
micro-service doesn't handle the error (minio not available yet) and restarts,
causing side effects that lead to 502 errors (connection refused from API
gateway because deployments micro-service is not yet up and running). While
effective, this fix is just a workaround: we wait 15 seconds before uploading
an artifact to the deployments service to be sure minio has completed its
startup process.

Changelog: None

Signed-off-by: Fabio Tranchitella <fabio@tranchitella.eu>