Remove --no-teardown feature of the test framework
It is unclear if anyone in the team is using it, and as it is it
introduces a hard dependency (the last one of this prototype!) on
common_docker file.

If we want to have such feature, it should be implemented in the
fixtures files by skipping the addfinalizer calls.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>