backend-tests: fix devauth test logic
If the device is in the preauthorization state and we are removing
authentication set with preauthorization status the devices should be
removed completely no matter if there are pending or rejected
authentication sets for the device or not.

The reason why the test was passing is that we were retrieving the device
which was just removed. It was possible with mongo 3.4 and the old
driver and it's still possible with mongo 3.6 and the new driver, but
the probability is lower.

changelog: none

Signed-off-by: Krzysztof Jaskiewicz <krzysiekjaskiewicz@gmail.com>