Port and re-enable test_os_ent_migration
Make the first fixture use the one defined in common_docker, and from
here pass it around to the rest of the fixtures and test logic instead
of calling global docker_compose funtions. Note that now the init data
is added to the fixture object iself.

Another imporant detail to notice is that this commit merges both test
cases into a single one and "donwgrades" the fixtures to function scope
to be compatible with the common_docker defined one.

We could still do better encapsulation here. We might come back to it if
the time permits.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>