QA-121: Unify the pytest environment variables passed to run.sh
Previously two environment variables were used, although the functionality was
complementary of eachother. The functionality was set, depending on whether or
not xdists was installed. If it was installed then either the XDISTS_ARGS
variables was passed to pytest. It xdist was not installed then pytest would be
given a --maxfail=1 parameter.

This commit unifies these options into one variable, the EXTRA_TEST_ARGS
variable, which is set according to the functionality described above.

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>