Optimize docker_compose_manager and rewrite test_os_ent_migration fixtures
Now that all related logic is in one place, we can tackle internal
optimizations and clean-ups in a low risk bases.

The main design change within this commit is to make each
DockerComposeNamespace child classes set the docker-compose files under
use in a private list at constructor time, so that the different methods
can access it directly and we avoid the parameters. The latter required
also a new method get_logs_of_service for one test that was still using
directly this method.

This in return let's us rewrite the fixtures in test_os_ent_migration to
directly manage it's required setups from the factory in a cleaner way.
An important related change to the test is removing the docker-compose
"template" for enterprise testing and instead use always the base one
with an emtpy DEFAULT_TOKEN when not present.

Also the fixtures are now restored to "class" scope and the test broken
down into two test cases (as it was originally).

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>